+++ datafusion/optimizer/src/push_down_limit.rs
@@ -427,9 +427,9 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  NoopPlan\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -449,9 +449,9 @@ mod test {
             .limit(10, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  NoopPlan\
-        \n    Limit: skip=Int64(0), fetch=Int64(1010)\
+        \n    Limit: skip=0, fetch=1010\
         \n      TableScan: test, fetch=1010";
 
         assert_optimized_plan_equal(plan, expected)
@@ -472,9 +472,9 @@ mod test {
             .limit(20, Some(500))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(30), fetch=Int64(500)\
+        let expected = "Limit: skip=30, fetch=500\
         \n  NoopPlan\
-        \n    Limit: skip=Int64(0), fetch=Int64(530)\
+        \n    Limit: skip=0, fetch=530\
         \n      TableScan: test, fetch=530";
 
         assert_optimized_plan_equal(plan, expected)
@@ -494,11 +494,11 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  NoopPlan\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -518,7 +518,7 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  NoLimitNoopPlan\
         \n    TableScan: test";
 
@@ -537,7 +537,7 @@ mod test {
         // Should push the limit down to table provider
         // When it has a select
         let expected = "Projection: test.a\
-        \n  Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n  Limit: skip=0, fetch=1000\
         \n    TableScan: test, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -555,7 +555,7 @@ mod test {
         // Should push down the smallest limit
         // Towards table scan
         // This rule doesn't replace multiple limits
-        let expected = "Limit: skip=Int64(0), fetch=Int64(10)\
+        let expected = "Limit: skip=0, fetch=10\
         \n  TableScan: test, fetch=10";
 
         assert_optimized_plan_equal(plan, expected)
@@ -571,7 +571,7 @@ mod test {
             .build()?;
 
         // Limit should *not* push down aggregate node
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Aggregate: groupBy=[[test.a]], aggr=[[max(test.b)]]\
         \n    TableScan: test";
 
@@ -588,11 +588,11 @@ mod test {
             .build()?;
 
         // Limit should push down through union
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Union\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -608,7 +608,7 @@ mod test {
             .build()?;
 
         // Should push down limit to sort
-        let expected = "Limit: skip=Int64(0), fetch=Int64(10)\
+        let expected = "Limit: skip=0, fetch=10\
         \n  Sort: test.a ASC NULLS LAST, fetch=10\
         \n    TableScan: test";
 
@@ -625,7 +625,7 @@ mod test {
             .build()?;
 
         // Should push down limit to sort
-        let expected = "Limit: skip=Int64(5), fetch=Int64(10)\
+        let expected = "Limit: skip=5, fetch=10\
         \n  Sort: test.a ASC NULLS LAST, fetch=15\
         \n    TableScan: test";
 
@@ -643,9 +643,9 @@ mod test {
             .build()?;
 
         // Limit should use deeper LIMIT 1000, but Limit 10 shouldn't push down aggregation
-        let expected = "Limit: skip=Int64(0), fetch=Int64(10)\
+        let expected = "Limit: skip=0, fetch=10\
         \n  Aggregate: groupBy=[[test.a]], aggr=[[max(test.b)]]\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -660,7 +660,7 @@ mod test {
 
         // Should not push any limit down to table provider
         // When it has a select
-        let expected = "Limit: skip=Int64(10), fetch=None\
+        let expected = "Limit: skip=10, fetch=None\
         \n  TableScan: test";
 
         assert_optimized_plan_equal(plan, expected)
@@ -678,7 +678,7 @@ mod test {
         // Should push the limit down to table provider
         // When it has a select
         let expected = "Projection: test.a\
-        \n  Limit: skip=Int64(10), fetch=Int64(1000)\
+        \n  Limit: skip=10, fetch=1000\
         \n    TableScan: test, fetch=1010";
 
         assert_optimized_plan_equal(plan, expected)
@@ -695,7 +695,7 @@ mod test {
             .build()?;
 
         let expected = "Projection: test.a\
-        \n  Limit: skip=Int64(10), fetch=Int64(990)\
+        \n  Limit: skip=10, fetch=990\
         \n    TableScan: test, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -712,7 +712,7 @@ mod test {
             .build()?;
 
         let expected = "Projection: test.a\
-        \n  Limit: skip=Int64(10), fetch=Int64(1000)\
+        \n  Limit: skip=10, fetch=1000\
         \n    TableScan: test, fetch=1010";
 
         assert_optimized_plan_equal(plan, expected)
@@ -728,7 +728,7 @@ mod test {
             .limit(0, Some(10))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(10), fetch=Int64(10)\
+        let expected = "Limit: skip=10, fetch=10\
         \n  TableScan: test, fetch=20";
 
         assert_optimized_plan_equal(plan, expected)
@@ -744,7 +744,7 @@ mod test {
             .build()?;
 
         // Limit should *not* push down aggregate node
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  Aggregate: groupBy=[[test.a]], aggr=[[max(test.b)]]\
         \n    TableScan: test";
 
@@ -761,11 +761,11 @@ mod test {
             .build()?;
 
         // Limit should push down through union
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  Union\
-        \n    Limit: skip=Int64(0), fetch=Int64(1010)\
+        \n    Limit: skip=0, fetch=1010\
         \n      TableScan: test, fetch=1010\
-        \n    Limit: skip=Int64(0), fetch=Int64(1010)\
+        \n    Limit: skip=0, fetch=1010\
         \n      TableScan: test, fetch=1010";
 
         assert_optimized_plan_equal(plan, expected)
@@ -787,7 +787,7 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in Join
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  Inner Join: test.a = test2.a\
         \n    TableScan: test\
         \n    TableScan: test2";
@@ -811,7 +811,7 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in Join
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  Inner Join: test.a = test2.a\
         \n    TableScan: test\
         \n    TableScan: test2";
@@ -836,7 +836,7 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in sub_query
-        let expected = "Limit: skip=Int64(10), fetch=Int64(100)\
+        let expected = "Limit: skip=10, fetch=100\
         \n  Filter: EXISTS (<subquery>)\
         \n    Subquery:\
         \n      Filter: test1.a = test1.a\
@@ -865,7 +865,7 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in sub_query
-        let expected = "Limit: skip=Int64(10), fetch=Int64(100)\
+        let expected = "Limit: skip=10, fetch=100\
         \n  Filter: EXISTS (<subquery>)\
         \n    Subquery:\
         \n      Filter: test1.a = test1.a\
@@ -893,11 +893,11 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Left Join: \
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)?;
@@ -912,11 +912,11 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Right Join: \
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)?;
@@ -931,11 +931,11 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Full Join: \
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)?;
@@ -950,9 +950,9 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  LeftSemi Join: \
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
         \n    TableScan: test2";
 
@@ -968,9 +968,9 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  LeftAnti Join: \
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
         \n    TableScan: test2";
 
@@ -986,10 +986,10 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  RightSemi Join: \
         \n    TableScan: test\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)?;
@@ -1004,10 +1004,10 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  RightAnti Join: \
         \n    TableScan: test\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1029,9 +1029,9 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in Join
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Left Join: test.a = test2.a\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
         \n    TableScan: test2";
 
@@ -1054,9 +1054,9 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in Join
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  Left Join: test.a = test2.a\
-        \n    Limit: skip=Int64(0), fetch=Int64(1010)\
+        \n    Limit: skip=0, fetch=1010\
         \n      TableScan: test, fetch=1010\
         \n    TableScan: test2";
 
@@ -1079,10 +1079,10 @@ mod test {
             .build()?;
 
         // Limit pushdown Not supported in Join
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  Right Join: test.a = test2.a\
         \n    TableScan: test\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1104,10 +1104,10 @@ mod test {
             .build()?;
 
         // Limit pushdown with offset supported in right outer join
-        let expected = "Limit: skip=Int64(10), fetch=Int64(1000)\
+        let expected = "Limit: skip=10, fetch=1000\
         \n  Right Join: test.a = test2.a\
         \n    TableScan: test\
-        \n    Limit: skip=Int64(0), fetch=Int64(1010)\
+        \n    Limit: skip=0, fetch=1010\
         \n      TableScan: test2, fetch=1010";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1123,11 +1123,11 @@ mod test {
             .limit(0, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(0), fetch=Int64(1000)\
+        let expected = "Limit: skip=0, fetch=1000\
         \n  CrossJoin:\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test, fetch=1000\
-        \n    Limit: skip=Int64(0), fetch=Int64(1000)\
+        \n    Limit: skip=0, fetch=1000\
         \n      TableScan: test2, fetch=1000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1143,11 +1143,11 @@ mod test {
             .limit(1000, Some(1000))?
             .build()?;
 
-        let expected = "Limit: skip=Int64(1000), fetch=Int64(1000)\
+        let expected = "Limit: skip=1000, fetch=1000\
         \n  CrossJoin:\
-        \n    Limit: skip=Int64(0), fetch=Int64(2000)\
+        \n    Limit: skip=0, fetch=2000\
         \n      TableScan: test, fetch=2000\
-        \n    Limit: skip=Int64(0), fetch=Int64(2000)\
+        \n    Limit: skip=0, fetch=2000\
         \n      TableScan: test2, fetch=2000";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1162,7 +1162,7 @@ mod test {
             .limit(1000, None)?
             .build()?;
 
-        let expected = "Limit: skip=Int64(1000), fetch=Int64(0)\
+        let expected = "Limit: skip=1000, fetch=0\
         \n  TableScan: test, fetch=0";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1177,7 +1177,7 @@ mod test {
             .limit(1000, None)?
             .build()?;
 
-        let expected = "Limit: skip=Int64(1000), fetch=Int64(0)\
+        let expected = "Limit: skip=1000, fetch=0\
         \n  TableScan: test, fetch=0";
 
         assert_optimized_plan_equal(plan, expected)
@@ -1194,7 +1194,7 @@ mod test {
             .build()?;
 
         let expected = "SubqueryAlias: a\
-        \n  Limit: skip=Int64(1000), fetch=Int64(0)\
+        \n  Limit: skip=1000, fetch=0\
         \n    TableScan: test, fetch=0";
 
         assert_optimized_plan_equal(plan, expected)
